db: remove pointless documentation
db-doc has an installed footprint of 94MB, but 60MB of that is the documentation
for the Java and C# bindings which are not part of this recipe.  Remove them and
the for-print PDF manuals to massively reduce the footprint of db-doc.

Also improve the fix for the documentation install path, and put the
documentation under ${docdir}/db instead of just ${docdir} (which is
/usr/share/doc by default).

(From OE-core rev: 89e9d8fc7f99fb6346dc035e51abbc942011d9b5)

Signed-off-by: Ross Burton <ross.burton@intel.com>